[tt-crank] feat(torch): input tensor handling utilities
Torch dispatch will call our kernels even when only one input tensor is
on our device. This means we need to convert all cpu input tensors to
our tensors.

Additionally, we need to handle different dtype scenarios:
1. input tensor elements can be 64-bit wide - this is not supported on our
   hardware currently. `tt-mlir` runtime downcasts the 64-bit wide types
   to 32-bit implicitly. When `.cpu()` is called on our tensors we will
   restore them back to 64-bit - if needed.
2. input tensors can be of different types - we use torch utilities to
   figure out what should the resulting dtype be and we will insert
   typecasts to inputs which need them (in the TTIR).

Generic test for dtype consistency is added - each additional kernel can
be added there accordingly.

`tt-mlir` is bumped to a version that handles casting + allocating host
buffer in `createOwnedHostTensor()`.

`copy_from` extended to handle `tensor.to(different_dtype)`.

Broadcasting support is implemented for the `add` kernel, as well.

Test infra updated with `device_type` fixture - so that we can control
the tests depending on if we're running on sim or on real device - dtype
tests are sensitive to that, since the sim fails on some dtypes.

+++ tt-crank/tests/python/op_tests/test_tensor.py
@@ -38,9 +38,11 @@ def test_empty_default_dtype_is_float() -> None:
 
 
 def test_empty_unsupported_dtype_rejected() -> None:
-    # to_runtime_dtype maps only bf16/f32/i32; f64 hits its default TORCH_CHECK.
+    # to_runtime_dtype covers the common scalar dtypes (bf16/f16/f32/f64,
+    # i32/i64, bool, u8); complex64 has no runtime mapping and hits the
+    # default TORCH_CHECK.
     with pytest.raises(RuntimeError, match="unsupported torch dtype"):
-        torch.empty((32, 32), device="tt", dtype=torch.float64)
+        torch.empty((32, 32), device="tt", dtype=torch.complex64)
 
 
 # -----------------------------------------------------------------------------
